requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
gunicorn>=20.1
djangorestframework